Predict gender of a user in ``TelegramAPI`` responses Remove hard coded title as `sir` and set as env var Remove hard coded name as Vignesh and set as env var Speak alarm deletion confirmation instead of printing on the screen

+++ executors/conditions.py
@@ -108,7 +108,8 @@ def conditions(converted: str, should_return: bool = False) -> bool:
             any(word in converted_lower for word in todo_checks):
         add_todo()
 
-    elif any(word in converted_lower for word in keywords.delete_todo):
+    elif any(word in converted_lower for word in keywords.delete_todo) and \
+            any(word in converted_lower for word in todo_checks):
         delete_todo()
 
     elif any(word in converted_lower for word in keywords.distance) and \
